Android example: Rework the page list "current page" implementation.
When the page list is showing on the right, we attempt to keep the content on the left
and the list on the ritgh in synch.

If you tap a page on the right, it gets highlighted and the content on the left scrolls
to show that page.

If you scroll the contents on the left, we pick a page in the middle, call that
"current", and then scroll the list on the right to show it.  We do this only
if being scrolled by the user (not when we're auto-scrolling for some other reason).

Fix the drawing of the border; it should be clipped when not fully
in view.