Eliminate jitter when pinch-zooming a document with many pages.
In each layout pass, we size and position each page.  This was done by calculating the
scaled size of each page as integers, then using those values in the layout calculation.
This would accumulate rounding errors as you scroll further down a long document.
Let's say you're on page 50.  A small change in the scale value could produce an extra
50 pixels in the layout.

Now we first do the layout calculation using the pages' natural sizes, which are
all fixed.  Then we apply the scale factor.