Remove loading of pages in the background; instead use lazy loading.
DocPageView now loads each muPDF page when it's first rendered.

For the purpose of computing layout, we assume that the page
is 8.5 x 11 inches until it is loaded, then we use the page's real size.

In this commit we've also removed a layout optimization that
was interfering with the search function.